chore(decodeCall): docs and move interface
